basic player movement within the tile map added
