+++ MazeSpooky/ProjectSettings/TagManager.asset
@@ -41,3 +41,9 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
+  - name: Player
+    uniqueID: 2988826461
+    locked: 0
+  - name: Background
+    uniqueID: 2178293549
+    locked: 0
